+++ media_platform/service/file_service/upload_file_response.py
@@ -1,4 +1,4 @@
-from media_platform.lang.serializable_deserializable import Deserializable
+from media_platform.lang.serialization import Deserializable
 from media_platform.service.file_descriptor import FileDescriptor
