+++ media_platform/service/file_service/upload_url.py
@@ -1,4 +1,4 @@
-from media_platform.lang.serializable_deserializable import Deserializable
+from media_platform.lang.serialization import Deserializable
 
 
 class UploadUrl(Deserializable):
