+++ media_platform/service/rest_result.py
@@ -1,5 +1,5 @@
 from media_platform.exception.media_platform_exception import MediaPlatformException
-from media_platform.lang.serializable_deserializable import Serializable, Deserializable
+from media_platform.lang.serialization import Serializable, Deserializable
 
 
 class RestResult(Serializable, Deserializable):
